refactor(agent): extract McpHttpTransport; Stop now joins the accept thread and drains workers
Extract the HTTP listener lifecycle out of the AgentServer monolith into an
instance McpHttpTransport (#215) taking its settings accessor and JSON-RPC
dispatch via constructor. The pure GateHttpRequest (#143) and the loopback
bind canonicalization (#152) move with it, unchanged.

Fixes on the way:
- Stop() closes the listener, JOINS the accept thread (bounded), and DRAINS
  the in-flight worker pool (bounded) so a Settings-dialog Stop/Start can no
  longer overlap old workers with a new listener.
- The HttpDispatchOverride static test seam is deleted; tests construct their
  own transport with an injected dispatch delegate instead.

AgentServer keeps StartHttp/StopHttp/IsHttpListening as thin facade members
(MainWindow's status strip, #211, keeps compiling against them).

Part of #215.
